The woman who has no imagination has no wings

+++ 0x0B-malloc_free/1-strdup.c
@@ -22,7 +22,7 @@ char *_strdup(char *str)
 	{
 	}
 
-	str_out = (char *)malloc(sizeof(char) * (unsigned long long)(i + 1));
+	str_out = (char *)malloc(sizeof(char) * (i + 1));
 
 	if (str_out == 0)
 		return (0);
